requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
lxml==5.1.0
packaging==23.2
playwright==1.41.1
pluggy==1.4.0
pyee==11.0.1
pytest==7.4.4
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-slugify==8.0.2
requests==2.31.0
soupsieve==2.5
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
